ivanmaldonado-2-api-planner-11f1ea10fd1c/requirements.txt
asgiref==3.7.2
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
